add run summary class for generating Run QC metrics
